+++ fabric/src/main/java/net/mehvahdjukaar/polytone/mixins/fabric/FogRendererMixin2.java
@@ -1,8 +1,7 @@
 package net.mehvahdjukaar.polytone.mixins.fabric;
 
 import com.llamalad7.mixinextras.sugar.Local;
-import com.mojang.blaze3d.systems.RenderSystem;
-import net.mehvahdjukaar.polytone.Polytone;
+import net.mehvahdjukaar.polytone.utils.FogManager;
 import net.minecraft.client.Camera;
 import net.minecraft.client.renderer.FogParameters;
 import net.minecraft.client.renderer.FogRenderer;
@@ -11,7 +10,6 @@ import org.joml.Vector4f;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(FogRenderer.class)
@@ -20,13 +18,21 @@ public abstract class FogRendererMixin2 {
     @Inject(method = "setupFog", at = @At(value = "TAIL"), cancellable = true)
     private static void polytone$modifyFogShape(Camera camera, FogRenderer.FogMode fogMode, Vector4f vector4f, float f, boolean bl, float g, CallbackInfoReturnable<FogParameters> cir, @Local FogType fogType) {
         if (fogMode == FogRenderer.FogMode.FOG_TERRAIN && fogType == FogType.NONE) {
-            var newFog = Polytone.BIOME_MODIFIERS.modifyFogParameters(
+            FogManager.FogState newFog = FogManager.modifyBiomeFog(
                     cir.getReturnValue().start(), cir.getReturnValue().end());
             if (newFog != null) {
                 FogParameters old = cir.getReturnValue();
-                cir.setReturnValue(new FogParameters(newFog.x, newFog.y, old.shape(), old.red(), old.green(), old.blue(), old.alpha()));
+                cir.setReturnValue(new FogParameters(newFog.start(), newFog.end(), old.shape(), old.red(), old.green(), old.blue(), old.alpha()));
             }
 
+        }
+        if (fogMode == FogRenderer.FogMode.FOG_TERRAIN && (fogType == FogType.WATER || fogType == FogType.LAVA)) {
+            FogManager.FogState newFog = FogManager.modifyFluidFog(
+                    cir.getReturnValue().start(), cir.getReturnValue().end(), null, null);
+            if (newFog != null) {
+                FogParameters old = cir.getReturnValue();
+                cir.setReturnValue(new FogParameters(newFog.start(), newFog.end(), old.shape(), old.red(), old.green(), old.blue(), old.alpha()));
+            }
         }
     }